+++ highs/ipm/hipo/factorhighs/Symbolic.h
@@ -21,7 +21,7 @@ struct TreeSchedule {
   std::vector<Int> task_parent;
 
   void clear();
-  Int count() const { return task_parent.size(); }
+  Int count() const { return static_cast<Int>(task_parent.size()); }
 };
 
 // Symbolic factorisation object
